fix(stark-core): avoid circular dependencies by moving state names in a constants file
