+++ packages/stark-core/src/modules/session.ts
@@ -1,5 +1,6 @@
 export * from "./session/actions";
 export * from "./session/components";
+export * from "./session/constants";
 export * from "./session/entities";
 export * from "./session/reducers";
 export * from "./session/services";
